requirements.txt
numpy
Pillow
pybind11
opencv-python-headless
face_recognition
facenet-pytorch
torch
torchvision
fastapi
uvicorn[standard]
python-multipart
